refactor(cli): split docs-index into its three concerns
577 lines held a YAML-subset parser, a markdown link scanner, and the
doc index itself. The cut is syntax versus semantics:

- docs-frontmatter.ts — the YAML subset (quoting, inline and block
  collections, mappings). Knows nothing about OKF fields.
- docs-links.ts — markdown link scanning (balanced parens, titles,
  angle-bracket destinations, escapes, reference definitions), plus
  localLinkTarget, the one shared bundle-local decision.
- docs-index.ts — DocRef, scanDocs, the OKF accessors, the entity
  index, and @docs tags. Stays the facade: it re-exports the parsers,
  so all seven importers and the public index are unchanged.

Tests split along the same lines. No behavior change.
